graphQL/requirements.txt
flask==3.0.0
flask-cors
graphene==3.3
graphql-core==3.2.3
requests==2.31.0
